pyproject.toml
[tool.poetry]
name = "sistema-recomendacao-de-produtos-por-frequencia"
version = "0.1.0"
description = ""
authors = ["bruno <b_linhares@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
streamlit = "^1.35.0"
altair = "^5.3.0"
cx-freeze = "^7.1.0"
pyinstaller = "^6.7.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
